air_automation/email_module/requirements.txt
openai==0.28
faiss-cpu
numpy
python-dotenv
torch
sympy
sentence_transformers
langchain
langchain-community
